artik/onboarding: enable HTTPs on the REST API
This patch changes the webserver to expose over HTTPs instead of HTTP
for channel encryption and avoid exposing sensitive information.

  - The server certificate is self-signed, clients need to have their
security policy configured accordingly to allow connecting to the server.
  - The the stack size of the threads that launch the web
server has been increased, as HTTPs initialization  requires more of it.
  - The mDNS service is now started after the web server, to ensure the
client that the REST API is ready to be accessed as soon as the client
retrieves the IP address of the device after mDNS discovery.
  - Increase version of the onboarding service to 1.6.0.

Change-Id: Ie677d38320338d1aedacc8cf04de8712c69184d9
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>